# fix: avoid-driver-context-switch